[x86][pti] Implement trampoline stacks
These are ministacks used for entering the kernel from usermode.  They
will be among the limited number of kstacks mapped into the uPML4 (the
only other exception being interrupt-stack-switching stacks for NMIs,
MCEs, and double faults).

These stacks are used only on entry into kernelmode (not exit), and
will be used to perform, in order: the gs swap, the PML4 swap, and the
switch to the thread's real kernel stack (including copying any
interrupt frame).  The stacks must be large enough to store an entire
interrupt frame (56 bytes for us) and scratch space for performing the above
operations.

NOTE: This implementation breaks usermode ASLR.  An attacker exploiting
Meltdown can read usermode instruction pointers and stack pointers off
of the percpu trampoline stacks.  Additionally, during syscalls the
usermode stack pointer is written to the leaked portion of the percpu
structure in gs.

Change-Id: I72f90d6e4dbdcdcde0e7c8c01524fad447c2cd7f