[kernel][ktrace] Add StringRef type to handle tracer strings.
Add the StringRef type to handle internalized strings with ids for
more generalized support of the ktrace subsystem. This is intented
to replace the current method of defining tracer strings in do-
while-false blocks that is incompatible with RAII and other scope-
aware tracing.

Due to a limitation in how section attributes are currently
handled, this implementation uses a runtime lock-free, wait-free
linked list to register strings in the string table. This can be
removed in favor of a linker script-defined array once Clang
and/or GCC have the appropriate support.

Test: Manual testing in a subsequent integration patch.
Change-Id: Ie529ca37509adc795b8ae5426c4c34250c1eacb6